Tiny change of architecture
in script.js : loaders functions were incapsulated in one function with
additional parameter 'params', that passes data to GET-query. It can be
real object with path or in can be null for ROOT directory.
That helped to divide logic in Controllers: there are method Get() for
ROOT value, and method Get(string path) for other pathes.